+++ app/gwells/authentication.py
@@ -16,7 +16,7 @@ from django.contrib.auth import get_user_model
 from rest_framework import exceptions
 from rest_framework_jwt.authentication import JSONWebTokenAuthentication
 from gwells.models.Profile import Profile
-from gwells.permissions import roles_to_groups
+from gwells.roles import roles_to_groups
 
 
 class JwtOidcAuthentication(JSONWebTokenAuthentication):
